fix: change telegram_chat_id column type to BigInteger and update database initialization

+++ app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, DateTime, Integer, String, false
+from sqlalchemy import BigInteger, Boolean, Column, DateTime, Integer, String, false
 from sqlalchemy.orm import relationship
 
 from app.models.base import Base, utcnow
@@ -8,7 +8,7 @@ class User(Base):
     __tablename__ = "users"
 
     id = Column(Integer, primary_key=True, index=True)
-    telegram_chat_id = Column(Integer, unique=True, nullable=False, index=True)
+    telegram_chat_id = Column(BigInteger, unique=True, nullable=False, index=True)
     telegram_username = Column(String(100), nullable=True)
     display_name = Column(String(200), nullable=True)
     is_admin = Column(Boolean, default=False, server_default=false(), nullable=False)
